Added: abstraction in util file
